LPS-34802 Moving ExportImportUtil functionality to ExportImportImpl to avoid confusion and adding new method for getting proper physical LAR file
